+++ src/DynamoCore/Models/NodeLoaders/ZeroTouchNodeLoader.cs
@@ -3,7 +3,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Xml;
-using DSCoreNodesUI;
 using Dynamo.DSEngine;
 using Dynamo.Nodes;
 using Dynamo.Utilities;
