+++ test/DynamoCoreTests/NodeMigrationTests.cs
@@ -1919,7 +1919,7 @@ namespace Dynamo.Tests
             Assert.AreEqual(5, workspace.Connectors.Count());
 
             // check that no nodes are migrated to dummy nodes
-            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is DSCoreNodesUI.DummyNode));
+            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is Dynamo.Nodes.DummyNode));
 
             // check that the node is migrated to a DSFunction nicknamed "ReferencePoint.ByPoint"
             StringAssert.Contains("Reference", workspace.NodeFromWorkspace<DSFunction>(
@@ -1938,7 +1938,7 @@ namespace Dynamo.Tests
             Assert.AreEqual(5, workspace.Connectors.Count());
 
             // check that no nodes are migrated to dummy nodes
-            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is DSCoreNodesUI.DummyNode));
+            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is Dynamo.Nodes.DummyNode));
 
             // check that the node is migrated to a DSFunction nicknamed "FamilyInstance.ByPoint"
             StringAssert.Contains("Instance", workspace.NodeFromWorkspace<DSFunction>(
@@ -1957,7 +1957,7 @@ namespace Dynamo.Tests
             Assert.AreEqual(5, workspace.Connectors.Count());
 
             // check that no nodes are migrated to dummy nodes
-            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is DSCoreNodesUI.DummyNode));
+            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is Dynamo.Nodes.DummyNode));
 
             // check that the node is migrated to a DSFunction nicknamed "ModelCurve.ByCurve"
             StringAssert.Contains("Model", workspace.NodeFromWorkspace<DSFunction>(
@@ -1975,7 +1975,7 @@ namespace Dynamo.Tests
             Assert.AreEqual(6, workspace.Connectors.Count());
 
             // check that no nodes are migrated to dummy nodes
-            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is DSCoreNodesUI.DummyNode));
+            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is Dynamo.Nodes.DummyNode));
 
             // check that the node is migrated to a PythonNode which retains the old script
             StringAssert.Contains("OUT = OUT", workspace.NodeFromWorkspace<DSIronPythonNode.PythonNode>(
@@ -1993,7 +1993,7 @@ namespace Dynamo.Tests
             Assert.AreEqual(6, workspace.Connectors.Count());
 
             // check that no nodes are migrated to dummy nodes
-            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is DSCoreNodesUI.DummyNode));
+            Assert.AreEqual(0, workspace.Nodes.AsQueryable().Count(x => x is Dynamo.Nodes.DummyNode));
 
             // check that some of the nodes are Excel nodes
             Assert.AreEqual(4, workspace.Nodes.AsQueryable().Count(x => x.NickName.Contains("Excel")));
@@ -2100,7 +2100,7 @@ namespace Dynamo.Tests
             int unresolvedNodeCount = 0;
             string str = "\n";
 
-            foreach (var node in nodes.OfType<DSCoreNodesUI.DummyNode>())
+            foreach (var node in nodes.OfType<Dynamo.Nodes.DummyNode>())
             {
                 if (node.NodeNature == DummyNode.Nature.Unresolved)
                 {
