+++ test/DynamoCoreWpfTests/RecordedTests.cs
@@ -2565,8 +2565,8 @@ namespace DynamoCoreWpfTests
                     var node1 = GetNode("37da4958-1b88-408b-b09d-3deba0ba3835");
                     var node2 = GetNode("b12ce9c8-8c23-43c4-987d-759c6f623998");
 
-                    Assert.NotNull(node1 as DSCoreNodesUI.DummyNode);
-                    Assert.NotNull(node2 as DSCoreNodesUI.DummyNode);
+                    Assert.NotNull(node1 as Dynamo.Nodes.DummyNode);
+                    Assert.NotNull(node2 as Dynamo.Nodes.DummyNode);
                 }
                 else if (commandTag == "Delete1")
                 {
