+++ test/DynamoCoreWpfTests/SerializationTests.cs
@@ -335,7 +335,7 @@ namespace DynamoCoreWpfTests
             ViewModel.OpenCommand.Execute(Path.Combine(folder, "DummyNodeSample.dyn"));
 
             var workspace = ViewModel.Model.CurrentWorkspace;
-            var dummyNode = workspace.NodeFromWorkspace<DSCoreNodesUI.DummyNode>(
+            var dummyNode = workspace.NodeFromWorkspace<Dynamo.Nodes.DummyNode>(
                 Guid.Parse("37bffbb9-3438-4c6c-81d6-7b41b5fb5b87"));
 
             Assert.IsNotNull(dummyNode);
@@ -358,7 +358,7 @@ namespace DynamoCoreWpfTests
             ViewModel.OpenCommand.Execute(Path.Combine(folder, "DummyNodeSample.dyn"));
 
             var workspace = ViewModel.Model.CurrentWorkspace;
-            var dummyNode = workspace.NodeFromWorkspace<DSCoreNodesUI.DummyNode>(
+            var dummyNode = workspace.NodeFromWorkspace<Dynamo.Nodes.DummyNode>(
                 Guid.Parse("37bffbb9-3438-4c6c-81d6-7b41b5fb5b87"));
 
             Assert.IsNotNull(dummyNode);
@@ -384,7 +384,7 @@ namespace DynamoCoreWpfTests
             ViewModel.OpenCommand.Execute(Path.Combine(folder, "dummyNode.dyn"));
 
             var workspace = ViewModel.Model.CurrentWorkspace;
-            var dummyNode = workspace.Nodes.OfType<DSCoreNodesUI.DummyNode>().FirstOrDefault();
+            var dummyNode = workspace.Nodes.OfType<Dynamo.Nodes.DummyNode>().FirstOrDefault();
 
             Assert.IsNotNull(dummyNode);
             var xmlDocument = new XmlDocument();
